Reduce code duplication in tests
- Moved common utility code to utils folder
- Specific code is present in the utils.py file in the test folders
- control_tests.py is renamed to config.py
