+++ tests/atlas/xaod/test_xaod_aggragate.py
@@ -1,14 +1,15 @@
 # Test out various things connected to the Aggregate call.
 # That code is more complex than I'd like it!
-from tests.atlas.xaod.utils_for_testing import (dataset_for_testing,
-                                                find_line_numbers_with,
-                                                find_line_with,
-                                                find_next_closing_bracket,
-                                                get_lines_of_code, print_lines, find_open_blocks)
-
+from tests.atlas.xaod.utils import (atlas_xaod_dataset,
+                                    get_lines_of_code,
+                                    print_lines,
+                                    find_line_numbers_with,
+                                    find_line_with,
+                                    find_next_closing_bracket,
+                                    find_open_blocks)
 
 def test_tree_name():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select("lambda e: e.Jets('AntiKt4EMTopoJets').Select(lambda j: j.pt()/1000).Sum()") \
         .AsROOTTTree('junk.root', 'analysis', ['fork']) \
         .value()
@@ -19,7 +20,7 @@ def test_tree_name():
 
 
 def test_Aggregate_not_initial_const_SUM():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select("lambda e: e.Jets('AntiKt4EMTopoJets').Select(lambda j: j.pt()/1000).Sum()") \
         .value()
     lines = get_lines_of_code(r)
@@ -29,7 +30,7 @@ def test_Aggregate_not_initial_const_SUM():
 
 
 def test_Aggregate_uses_floats_for_float_sum():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select("lambda e: e.Jets('AntiKt4EMTopoJets').Select(lambda j: j.pt()/1000).Sum()") \
         .value()
     lines = get_lines_of_code(r)
@@ -39,7 +40,7 @@ def test_Aggregate_uses_floats_for_float_sum():
 
 
 def test_count_after_single_sequence():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").Select(lambda j: j.pt()).Count()') \
         .value()
     lines = get_lines_of_code(r)
@@ -54,7 +55,7 @@ def test_count_after_single_sequence():
 
 
 def test_count_after_single_sequence_with_filter():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").Select(lambda j: j.pt()).Where(lambda jpt: jpt>10.0).Count()') \
         .value()
     lines = get_lines_of_code(r)
@@ -69,7 +70,7 @@ def test_count_after_single_sequence_with_filter():
 
 
 def test_count_after_double_sequence():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").SelectMany(lambda j: e.Tracks("InnerTracks")).Count()') \
         .value()
     lines = get_lines_of_code(r)
@@ -84,7 +85,7 @@ def test_count_after_double_sequence():
 
 
 def test_count_after_single_sequence_of_sequence():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").Select(lambda j: e.Tracks("InnerTracks")).Count()') \
         .value()
     lines = get_lines_of_code(r)
@@ -99,7 +100,7 @@ def test_count_after_single_sequence_of_sequence():
 
 
 def test_count_after_double_sequence_with_filter():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").SelectMany(lambda j: e.Tracks("InnerTracks").Where(lambda t: t.pt()>10.0)).Count()') \
         .value()
     lines = get_lines_of_code(r)
@@ -114,7 +115,7 @@ def test_count_after_double_sequence_with_filter():
 
 
 def test_count_after_single_sequence_of_sequence_unwound():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").Select(lambda j: e.Tracks("InnerTracks")).SelectMany(lambda ts: ts).Count()') \
         .value()
     lines = get_lines_of_code(r)
@@ -129,7 +130,7 @@ def test_count_after_single_sequence_of_sequence_unwound():
 
 
 def test_count_after_single_sequence_of_sequence_with_useless_where():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").Select(lambda j: e.Tracks("InnerTracks").Where(lambda pt: pt > 10.0)).Count()') \
         .value()
     lines = get_lines_of_code(r)
@@ -148,26 +149,26 @@ def test_count_after_single_sequence_of_sequence_with_useless_where():
 def test_first_can_be_iterable_after_where():
     # This was found while trying to generate a tuple for some training, below, simplified.
     # The problem was that First() always returned something you weren't allowed to iterate over. Which is not what we want here.
-    dataset_for_testing() \
+    atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").Select(lambda j: e.Tracks("InnerTracks").Where(lambda t: t.pt() > 1000.0)).First().Count()') \
         .value()
 
 
 def test_first_can_be_iterable():
     # Make sure a First() here gets called back correctly and generated.
-    dataset_for_testing() \
+    atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AllMyJets").Select(lambda j: e.Tracks("InnerTracks")).First().Count()') \
         .value()
 
 
 def test_Aggregate_per_jet():
-    dataset_for_testing() \
+    atlas_xaod_dataset() \
         .Select("lambda e: e.Jets('AntiKt4EMTopoJets').Select(lambda j: j.pt()).Count()") \
         .value()
 
 
 def test_Aggregate_per_jet_int():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select("lambda e: e.Jets('AntiKt4EMTopoJets').Select(lambda j: j.pt()).Count()") \
         .value()
 
@@ -178,7 +179,7 @@ def test_Aggregate_per_jet_int():
 
 
 def test_generate_Max():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select("lambda e: e.Jets('AntiKt4EMTopoJets').Select(lambda j: j.pt()).Max()") \
         .value()
     lines = get_lines_of_code(r)
@@ -187,7 +188,7 @@ def test_generate_Max():
 
 def test_First_selects_collection_count():
     # Make sure that we have the "First" predicate after if Where's if statement.
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AntiKt4EMTopoJets").Select(lambda j: e.Tracks("InDetTrackParticles")).First().Count()') \
         .value()
     lines = get_lines_of_code(r)
@@ -197,7 +198,7 @@ def test_First_selects_collection_count():
 
 
 def test_sequence_with_where_first():
-    r = dataset_for_testing() \
+    r = atlas_xaod_dataset() \
         .Select('lambda e: e.Jets("AntiKt4EMTopoJets").Select(lambda j: e.Tracks("InDetTrackParticles").Where(lambda t: t.pt() > 1000.0)).First().Count()') \
         .value()
     lines = get_lines_of_code(r)
