+++ tests/common/test_cpp_ast.py
@@ -2,11 +2,11 @@ import pytest
 
 from func_adl_xAOD.common.math_utils import DeltaR
 
-from tests.atlas.xaod.utils_for_testing import dataset_for_testing
+from tests.atlas.xaod.utils import atlas_xaod_dataset
 
 
 def test_deltaR_call():
-    r = dataset_for_testing().Select('lambda e: DeltaR(1.0, 1.0, 1.0, 1.0)').value()
+    r = atlas_xaod_dataset().Select('lambda e: DeltaR(1.0, 1.0, 1.0, 1.0)').value()
     vs = r.QueryVisitor._gc._class_vars
     assert 1 == len(vs)
     assert "double" == str(vs[0].cpp_type())
@@ -14,4 +14,4 @@ def test_deltaR_call():
 
 def test__bad_deltaR_call():
     with pytest.raises(ValueError):
-        dataset_for_testing().Select('lambda e: DeltaR(1.0, 1.0, 1.0)').value()
+        atlas_xaod_dataset().Select('lambda e: DeltaR(1.0, 1.0, 1.0)').value()
